challenges/5_Van_Halen_Radiation_Belt/3_Quantum/solver/requirements.txt
numpy==1.24.2
pycryptodomex==3.17
pyctf
